local-inference/requirements.txt
# OpenCLI Local Inference Dependencies
# Install: pip install -r requirements.txt

torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.27.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0
Pillow>=10.0.0
numpy>=1.24.0
huggingface-hub>=0.20.0
edge-tts>=6.1.0
realesrgan>=0.3.0
basicsr>=1.4.2
gfpgan>=1.3.8
controlnet-aux>=0.0.7
